requirements.txt
python-telegram-bot==12.0.0b1
beautifulsoup4
requests
PySocks
lxml
psycopg2-binary